+++ gamevault/ViewModels/SettingsViewModel.cs
@@ -181,7 +181,7 @@ namespace gamevault.ViewModels
         {
             get
             {
-                return "1.5.0";
+                return "1.6.0";
             }
         }