+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # GameVault App Changelog
 
+## 1.6.0
+### Changes
+- Faster Rawg search in the game view
+- More detailed error messages when downloading
+- Added support for more executable files
+- Added get random game button to library
+- Added new Exception Window where you can choose whether to send a crash report or not
+- Show certain data in the gameview only if they are available
+- Update Notification for new Versions in the non Microsoft Store Releases
+- Check for offline cache integrity
+- Uninstall games
+- Bug fix: Crash if you navigate to Settings/Data and never downloaded a game before
+- Bug fix: Image was requested although an invalid image id was specified
+- Bug fix: Game Title was also not displayed when the Release Date was not set
+- Bug fix: Could not navigate to errorlog on Microsoft Store version
+
 ## 1.5.0
 ### Changes
 - Reorganized settings tab